refactor(backend): move projects domain into farm/projects and retire farm/views.py
Move project, membership, invitation and agent-login views plus the
invitation email helper into a farm/projects domain package, and
VersionView into farm/common/views.py. Delete the now-empty
farm/views.py and farm/serializers.py monoliths, repoint
config/urls.py and the test mock patch target, and update comment
references in crops/ and farm/services. Route table unchanged.
