+++ backend/farm/projects/__init__.py
@@ -1,0 +1,1 @@
+"""Projects domain: projects, memberships, invitations, and agent login."""
